Get rid of explicit channel closure in BLE scan handler
I hit a rare edge case where `go-ble` invoked the advertisement
callback after `Scan()` returned and channel cleanup had happened,
causing the exporter to crash due to a write on a closed channel.

Closing channels is unnecessary unless the sender is using the closure
as a way to indicate the operation being done. In this case, `context.Context`
is used as the cancellation dispatcher, so it isn't really necessary
to close any channel. They will all be GC'd when the function ends.

For sanity and efficiency, check whether the context is done before
enqueueing a new advertisement anyway.